Fix 'DS structure' related clear and value deletion functions

+++ world-against-us/scripts/ClearDSListAndDeleteValues/ClearDSListAndDeleteValues.gml
@@ -2,12 +2,18 @@ function ClearDSListAndDeleteValues(_dsList, _valueType = undefined)
 {
 	if (!is_undefined(_dsList))
 	{
-		var listSize = ds_list_size(_dsList);
-		repeat(listSize)
+		if (ds_exists(_dsList, ds_type_list))
 		{
-			DeleteDSListValueByIndex(_dsList, 0, _valueType);
+			var listSize = ds_list_size(_dsList);
+			repeat(listSize)
+			{
+				DeleteDSListValueByIndex(_dsList, 0, _valueType);
+			}
+			ds_list_clear(_dsList);
+		} else {
+			// TODO: Generic error handling
+			show_debug_message("Unable to clear unknown DS list");
 		}
-		ds_list_clear(_dsList);
 	} else {
 		// TODO: Generic error handling
 		show_debug_message("Unable to clear undefined DS list");
